Move And matcher to separate file
